+++ src/redturtle/volto/restapi/serializer/summary.py
@@ -24,21 +24,38 @@ EMPTY_STRINGS = ["None"]
 @implementer(ISerializeToJsonSummary)
 @adapter(Interface, IRedturtleVoltoLayer)
 class DefaultJSONSummarySerializer(BaseSerializer):
-    def get_metadata_fields(self):
+    def __init__(self, context, request):
+        super().__init__(context, request)
+        self.force_all_metadata = False
+
+    def metadata_fields(self):
+        metadata_fields = super().metadata_fields()
+        if not self.force_all_metadata:
+            return metadata_fields
+        fields_cache = self.request.get("_summary_fields_cache", None)
+        if fields_cache is None:
+            catalog = api.portal.get_tool(name="portal_catalog")
+            fields_cache = set(catalog.schema()) | self.non_metadata_attributes
+            self.request.set("_summary_fields_cache", fields_cache)
+        return metadata_fields | fields_cache
+
+    @property
+    def show_all_metadata_fields(self):
         query = self.request.form
         if not query:
             # maybe its a POST request
             query = json_body(self.request)
-        return query.get("metadata_fields", []) or []
+        metadata_fields = query.get("metadata_fields", [])
+        return "_all" in metadata_fields or self.force_all_metadata
 
     def get_image_scales(self, data):
         """
         this is a backward compatibility for old volto templates that need
         a full image scales object
         """
-        metadata_fields = self.get_metadata_fields()
+        metadata_fields = self.metadata_fields()
 
-        if "_all" not in metadata_fields:
+        if "image" not in metadata_fields and not self.show_all_metadata_fields:
             return {}
         if data.get("image", None):
             # it's a fullobjects data, so we already have the infos
@@ -46,6 +63,7 @@ class DefaultJSONSummarySerializer(BaseSerializer):
         if not data.get("image_field", ""):
             return None
         scales = {}
+
         for name, actual_width, actual_height in get_scale_infos():
             scales[name] = {
                 "width": actual_width,
@@ -81,9 +99,10 @@ class DefaultJSONSummarySerializer(BaseSerializer):
                 return ""
         return ""
 
-    def __call__(self):
+    def __call__(self, force_all_metadata=False):
+        if force_all_metadata:
+            self.force_all_metadata = True
         data = super().__call__()
-        metadata_fields = self.get_metadata_fields()
         # return empty values if dates are not set:
         for k, v in data.items():
             if isinstance(v, str):
@@ -93,16 +112,16 @@ class DefaultJSONSummarySerializer(BaseSerializer):
                 # this is a Volto compatibility
                 data[k] = None
         scales = self.get_image_scales(data)
+
         if scales:
             data["image"] = {"scales": scales}
         if self.context.portal_type == "Link":
-            if "_all" in metadata_fields or "remoteUrl" in metadata_fields:
-                remote_url = self.get_remote_url()
-                # set twice because old templates can use both
-                if not remote_url:
-                    data["getRemoteUrl"] = ""
-                    data["remoteUrl"] = ""
-                else:
-                    data["remoteUrl"] = remote_url
-                    data["getRemoteUrl"] = remote_url
+            remote_url = self.get_remote_url()
+            # set twice because old templates can use both
+            if not remote_url:
+                data["getRemoteUrl"] = ""
+                data["remoteUrl"] = ""
+            else:
+                data["remoteUrl"] = remote_url
+                data["getRemoteUrl"] = remote_url
         return data